+++ src/decorators/validator/string/minStringSize.ts
@@ -3,7 +3,7 @@ import { NameValidator } from '../error/validationError';
 import { validate } from '../method';
 import { required } from '../parameter';
 import { messageMapper, Validator } from '../type';
-import { addNeededValidateParams, addValidatorForParams } from '../../utils/metadata';
+import { addNeededValidateParams, addHandlerForParams } from '../../utils/metadata';
 
 class MinStringSizeValidator extends BaseValidator implements Validator {
   protected name: NameValidator = NameValidator.MIN_SIZE;
@@ -29,7 +29,7 @@ function MinStringSize<T extends Error>(minSize: number, errorHandler?: T | stri
     validator.setErrorHandler(errorHandler);
     validator.setMinSize(minSize);
 
-    addValidatorForParams(target, propertyKey, validator);
+    addHandlerForParams(target, propertyKey, validator);
   };
 }