+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [1.2.2]
+
+- Add `Mapping` decorator to change field name
+- Fix process decorators
+- Add integration tests
+- Add more methods for logger
+
 # [1.2.1]
 
 - remove strict node engine